+++ lib/basic_utils.dart
@@ -1,29 +1,22 @@
 library basic_utils;
 
-import 'dart:convert';
-import 'dart:math';
+export "src/model/CountryCodeList.dart";
+export "src/model/Domain.dart";
+export "src/model/EmailAddress.dart";
+export "src/model/GtldList.dart";
+export "src/model/IdnCountryCodeList.dart";
+export "src/model/PublicSuffix.dart";
+export "src/model/LengthUnits.dart";
+export "src/model/exception/HttpResponseException.dart";
+export "src/model/RRecordType.dart";
+export "src/model/RRecord.dart";
+export "src/model/ResolveResponse.dart";
+export "src/model/HttpRequestReturnType.dart";
 
-import 'package:http/http.dart' as http;
-import 'package:http/http.dart';
-import 'package:logging/logging.dart';
-
-import "src/model/CountryCodeList.dart";
-import "src/model/Domain.dart";
-import "src/model/EmailAddress.dart";
-import "src/model/GtldList.dart";
-import "src/model/IdnCountryCodeList.dart";
-import "src/model/PublicSuffix.dart";
-import "src/model/LengthUnits.dart";
-import "src/model/exception/HttpResponseException.dart";
-import "src/model/RRecordType.dart";
-import "src/model/RRecord.dart";
-import "src/model/ResolveResponse.dart";
-import "src/model/HttpRequestReturnType.dart";
-
-part "src/DomainUtils.dart";
-part "src/EmailUtils.dart";
-part "src/StringUtils.dart";
-part "src/MathUtils.dart";
-part "src/HttpUtils.dart";
-part "src/DnsUtils.dart";
-part "src/SortUtils.dart";
+export "src/DomainUtils.dart";
+export "src/EmailUtils.dart";
+export "src/StringUtils.dart";
+export "src/MathUtils.dart";
+export "src/HttpUtils.dart";
+export "src/DnsUtils.dart";
+export "src/SortUtils.dart";